+++ datafusion/Cargo.toml
@@ -51,6 +51,7 @@ avro = ["avro-rs", "num-traits", "datafusion-common/avro"]
 
 [dependencies]
 datafusion-common = { path = "../datafusion-common", version = "6.0.0" }
+datafusion-expr = { path = "../datafusion-expr", version = "6.0.0" }
 ahash = { version = "0.7", default-features = false }
 hashbrown = { version = "0.12", features = ["raw"] }
 arrow = { version = "8.0.0", features = ["prettyprint"] }
